# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Music APIs",
    "version": "1.0.0",
    "description": "API documentation for musicApis project"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/songs/upload": {
      "post": {
        "summary": "Upload a song file",
        "tags": ["Songs"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["songFile", "name", "artist", "duration"],
                "properties": {
                  "songFile": {
                    "type": "string",
                    "format": "binary"
                  },
                  "name": {
                    "type": "string"
                  },
                  "artist": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "string"
                  },
                  "img": {
                    "type": "string"
                  },
                  "song": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Song uploaded successfully"
          }
        }
      }
    },
    "/songs": {
      "get": {
        "summary": "Get all songs",
        "tags": ["Songs"],
        "responses": {
          "200": {
            "description": "List of all songs"
          }
        }
      },
      "post": {
        "summary": "Create a new song (no file upload)",
        "tags": ["Songs"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "artist"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "artist": {
                    "type": "string"
                  },
                  "songFile": {
                    "type": "string"
                  },
                  "img": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "string"
                  },
                  "song": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Song created successfully"
          }
        }
      }
    },
    "/songs/{id}": {
      "get": {
        "summary": "Get a song by ID",
        "tags": ["Songs"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Song data"
          },
          "404": {
            "description": "Song not found"
          }
        }
      },
      "put": {
        "summary": "Update a song",
        "tags": ["Songs"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "artist": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Song updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a song",
        "tags": ["Songs"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Song deleted"
          }
        }
      }
    },
    "/songs/like/{id}": {
      "put": {
        "summary": "Like or unlike a song",
        "tags": ["Songs"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Like toggled"
          }
        }
      }
    },
    "/songs/like": {
      "get": {
        "summary": "Get liked songs of a user",
        "tags": ["Songs"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of liked songs"
          }
        }
      }
    },
    "/users": {
      "post": {
        "summary": "Create a new user",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created successfully"
          }
        }
      },
      "get": {
        "summary": "Get all users",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of users"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get a user by ID",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User data"
          }
        }
      },
      "put": {
        "summary": "Update user by ID",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete user by ID",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          }
        }
      }
    }
  },
...
}